+++ Sarah's Welt/src/world/creatures/Trex.java
@@ -29,7 +29,7 @@ public class Trex extends WalkingCreature {
 	}
 	
 	@Override
-	public void update(int dTime){
+	public void update(float dTime){
 		if(g){
 //			pos.y++;
 //			accelerateFromGround(new Point(0, 0.001f));
